% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianLaterality-package.R
\docType{package}
\name{BayesianLaterality-package}
\alias{BayesianLaterality}
\alias{BayesianLaterality-package}
\title{BayesianLaterality: Predict Brain Asymmetry Based on Handedness and Dichotic Listening}
\description{
Functional differences between the cerebral hemispheres are a fundamental characteristic of the human brain. Researchers interested in studying these differences often infer underlying hemispheric dominance for a certain function (e.g., language) from laterality indices calculated from observed performance or brain activation measures . However, any inference from observed measures to latent (unobserved) classes has to consider the prior probability of class membership in the population. The provided functions implement a Bayesian model for predicting hemispheric dominance from observed laterality indices (Sorensen and Westerhausen, Laterality: Asymmetries of Body, Brain and Cognition, 2020, \doi{10.1080/1357650X.2020.1769124}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LCBC-UiO/BayesianLaterality}
}

}
\author{
\strong{Maintainer}: Oystein Sorensen \email{oystein.sorensen.1985@gmail.com} (\href{https://orcid.org/0000-0003-0724-3542}{ORCID})

}
\keyword{internal}
