% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesianLaterality.R
\docType{package}
\name{BayesianLaterality}
\alias{BayesianLaterality}
\title{BayesianLaterality: Predict hemispheric dominance based on observed laterality measures.}
\description{
The BayesianLaterality package provides functionality for
predicting hemispheric dominance based on observed laterality measures,
using the Bayes classifier described in
\insertCite{Sorensen2020;textual}{BayesianLaterality}. At the moment,
the only function is \code{\link{predict_dominance}}.
}
\references{
\insertAllCited{}
}
