% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data1}
\alias{example_data1}
\title{Example dataset with a single measurement of three individuals.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 2 columns.
}
\usage{
example_data1
}
\description{
Example dataset with a single measurement of three individuals.
}
\keyword{datasets}
