% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data2}
\alias{example_data2}
\title{Example dataset with three measurements each on 100 individuals.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 300 rows and 4 columns.
}
\usage{
example_data2
}
\description{
Example dataset with three measurements each on 100 individuals.
}
\keyword{datasets}
