% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_asymmetry.R
\name{predict_asymmetry}
\alias{predict_asymmetry}
\title{Predict hemispheric dominance}
\usage{
predict_asymmetry(
  data,
  parameters = dplyr::tibble(dominance = rep(c("left", "right", "none"), each = 2),
    handedness = rep(c("left", "right"), 3), mean_li = c(10, 12, -24, -24, 0, 0), sd_li =
    rep(22, 6), prob_dominance = c(0.65, 0.87, 0.2, 0.04, 0.15, 0.09)),
  truncation = c(-100, 100),
  icc = NULL
)
}
\arguments{
\item{data}{Data frame with the following columns:
\itemize{
\item \code{listening}: Score between -100 and 100.
\item \code{handedness}: \code{"left"} for adextral (non-right-handed) and
      \code{"right"} for dextral (right-handed)
      }
 In addition, an optional column name \code{ID}
  can be provided, giving the subject ID. If a subject has multiple
  measurements, the posterior based on all measurements is provided. If the
  \code{ID} column is missing, each row is assumed to be measured on a
  separate subject.}

\item{parameters}{Data frame in which the first two columns specify combinations
of hemispheric dominance and handedness and the last three columns specify
the corresponding parameter values. In particular, the columns are defined as follows:
\itemize{
\item \code{dominance}: character specifying hemispheric dominance.
\item \code{handedness}: character specifying handedness.
\item \code{mean_li}: mean dichotic listening score.
\item \code{sd_li}: standard deviation of dichotic listening score.
\item \code{prob_dominance}: probability of hemispheric dominance given handedness.
}}

\item{truncation}{Numeric vector with two elements specifying the lower and upper
bounds for truncation of the normal distribution for dichotic listening scores.}

\item{icc}{Intraclass correlation for repeated measurements on the same individual.
If not specified, and if \code{data} contains repeated measurements, it is computed
from the data. Setting \code{icc=0} implies that repeated measurements on the same
individual are assumed to be conditionally independent given handedness and
hemispheric dominance.}
}
\value{
The probability of left or right hemispheric dominance in additional
  columns of \code{data}.
}
\description{
Predict hemispheric dominance
}
\examples{
## Simple test dataset
data <- data.frame(
          listening = c(-20, -23, -14),
          handedness = "left",
          stringsAsFactors = FALSE
         )
## Compute predictions
predict_asymmetry(data)

## More interesting example, with multiple measurements per individual.
library(dplyr); library(purrr); library(tidyr); library(truncnorm)
## First we sample test data
n <- 100 # number of individuals
reps <- 3 # number of measurements per individual
## The distribution of subject means has standard deviation 10, and the
## actual measurements for each subject are distributed with a standard
## deviation of 10 around this mean.
set.seed(234)
data <- tibble(
               ID = factor(1:n),
               subject_mean = rtruncnorm(n, a = 0, b = 100, mean = 10, sd = 10),
               handedness = "left") \%>\%
  mutate(
    listening = map(subject_mean, ~ rtruncnorm(reps, a = -100, b = 100,
                    mean = .x, sd = 10))
  ) \%>\%
  unnest(listening)

predict_asymmetry(data)


}
