% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_asymmetry.R
\name{predict_asymmetry}
\alias{predict_asymmetry}
\title{Predict brain asymmetry}
\usage{
predict_asymmetry(data, mu = c(10.43, -24.41, 11.52, -24.41),
  sigma = c(22.8, 28, 17, 28), rho = c(0.31, 0.115))
}
\arguments{
\item{data}{Data frame with the following columns:
\itemize{
\item \code{listening}: Score between -100 and 100.
\item \code{handedness}: \code{"L"} or \code{"A"} for adextral (non-right-handed) and
\code{"R"} or \code{"D"} for dextrals (right-handed)
}
In additional, an optional column name \code{ID} can be provided, giving the subject ID. If a
subject has multiple measurements, the posterior based on all measurements is provided.}

\item{mu}{Vector of mean dichotic listening scores.}

\item{sigma}{Vector of standard deviation of dichotic listening scores.}

\item{rho}{Probability of dextrality given left or right hemispheric dominance.}
}
\value{
The probability of left or right hemispheric dominance in additional columns of \code{data}.
}
\description{
Predict brain asymmetry
}
\examples{
library(asymm)
data <- data.frame(handedness = c("A", "D", "D"),
                   listening = c(-10, 20, 10),
                   stringsAsFactors = FALSE)

predict_asymmetry(data)


data <- data.frame(ID = c(1, 1, 2),
                   handedness = c("A", "A", "D"),
                   listening = c(-10, 20, 10),
                   stringsAsFactors = FALSE)

}
