% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_asymmetry.R
\name{predict_asymmetry}
\alias{predict_asymmetry}
\title{Predict brain asymmetry}
\usage{
predict_asymmetry(
  data,
  mu = c(10, -24, 12, -24),
  sigma = rep(22, 4),
  rho = c(0.2, 0.04),
  mu_sem = NULL
)
}
\arguments{
\item{data}{Data frame with the following columns:
\itemize{
\item \code{listening}: Score between -100 and 100.
\item \code{handedness}: \code{"L"} or \code{"A"} for adextral (non-right-handed) and
\code{"R"} or \code{"D"} for dextrals (right-handed)
}
In additional, an optional column name \code{ID} can be provided, giving the subject ID. If a
subject has multiple measurements, the posterior based on all measurements is provided. If the
\code{ID} column is missing, each row is assumed to be measured on a separate subject.}

\item{mu}{Vector of mean dichotic listening scores.}

\item{sigma}{Vector of standard deviation of dichotic listening scores.}

\item{rho}{Probability of dextrality given left or right hemispheric dominance.}

\item{mu_sem}{Vector representing the standard error of the mean of \code{mu}. If provided, a hierarchical model
is computed.}
}
\value{
The probability of left or right hemispheric dominance in additional columns of \code{data}.
}
\description{
Predict brain asymmetry
}
\examples{
library(asymm)

# Three observations of three individuals
data <- data.frame(handedness = c("A", "D", "D"),
                   listening = c(-10, 20, 10),
                   stringsAsFactors = FALSE)

predict_asymmetry(data)

# Take into account uncertainty in the parameters for mean speech laterlization scores.
predict_asymmetry(data, mu_sem = c(4.6, 7.0, 4.3, 7.0))

# Multiple observations
data <- data.frame(ID = c(1, 1, 2),
                   handedness = c("A", "A", "D"),
                   listening = c(-10, 20, 10),
                   stringsAsFactors = FALSE)

predict_asymmetry(data)

# Plot over a grid
grid <- expand.grid(listening = seq(from = -100, to = 100, by = 1), handedness = c("A", "D"))
grid$ID <- seq(1, nrow(grid), by = 1)
predictions <- merge(grid, predict_asymmetry(grid), by = "ID")

library(ggplot2)
ggplot(predictions, aes(x = listening, y = LeftDominance, group = handedness, color = handedness)) +
  geom_line() +
  geom_hline(yintercept = 0.05, color = "gray", lwd = 1) +
  geom_hline(yintercept = 0.95, color = "gray", lwd = 1) +
  xlab("Speech lateralization score") +
  ylab("Probability of left brain asymmetry") +
  theme(text = element_text(size = 16),
        legend.title = element_blank())

}
